Improve user experience

+++ BatchRenamerExtension/Libs/Renamer.cs
@@ -11,9 +11,9 @@ namespace BatchRenamerExtension
         public delegate void ShowError(string title, string message);
         public delegate bool AskForContinuation(string title, string message);
         public static Tuple<List<FileOrFolderPath>, bool> Rename(
-            PathContainer sources, 
-            PathContainer dests, 
-            ShowError ShowError, 
+            PathContainer sources,
+            PathContainer dests,
+            ShowError ShowError,
             AskForContinuation Ask)
         {
             var success = new List<FileOrFolderPath>();
@@ -27,34 +27,26 @@ namespace BatchRenamerExtension
             var renames = dests.Zip(sources, (f, s) => Tuple.Create(f, s))
                 .Where(x => x.Item1.CompletePath != x.Item2.CompletePath);
 
+            List<Tuple<FileOrFolderPath, FileOrFolderPath>> overwriting = new List<Tuple<FileOrFolderPath, FileOrFolderPath>>();
+            List<Tuple<FileOrFolderPath, FileOrFolderPath>> inUse = new List<Tuple<FileOrFolderPath, FileOrFolderPath>>();
+            List<Tuple<FileOrFolderPath, FileOrFolderPath>> invalidPath = new List<Tuple<FileOrFolderPath, FileOrFolderPath>>();
+            List<Tuple<FileOrFolderPath, FileOrFolderPath>> folderMerge = new List<Tuple<FileOrFolderPath, FileOrFolderPath>>();
+            List<Tuple<FileOrFolderPath, FileOrFolderPath>> ok = new List<Tuple<FileOrFolderPath, FileOrFolderPath>>();
             foreach (var p in renames)
             {
                 var source = p.Item2;
                 var dest = p.Item1;
                 try
                 {
-                    if (source.ExistsAsFile && dest.ExistsAsFile)
-                    { 
-                        if(!Ask("Overwriting", String.Format("Do you want to override {0} with {1}", source.CompletePath, dest.CompletePath)))
-                        {
-                            return Tuple.Create(success, false);
-                        }
-                    }
-                    if (source.ExistsAsFile && !dest.IsValidAsFile)
-                    {
-                        ShowError("Invalid file path", String.Format("The file {0} has an invalid path or name", dest.CompletePath));
-                        return Tuple.Create(success, false);
-                    }
-                    if (source.ExistsAsDirectory && dest.ExistsAsDirectory)
-                    {
-                        ShowError("Folder merge", String.Format("Folder merge is not supported: {0}", dest.CompletePath));
-                        return Tuple.Create(success, false);
-                    }
-                    if(source.ExistsAsDirectory && !dest.IsValidAsDirectory(source.CompletePath))
-                    {
-                        ShowError("Invalid folder path", String.Format("The folder {0} has an invalid path or name", dest.CompletePath));
-                        return Tuple.Create(success, false);
-                    }
+                    source.InvalidateCache();
+                    dest.InvalidateCache();
+                    if (source.ExistsAsFile && dest.ExistsAsFile) overwriting.Add(p);
+                    else if (source.ExistsAsFile && !dest.IsValidAsFile) invalidPath.Add(p);
+                    else if (source.ExistsAsDirectory && dest.ExistsAsDirectory) folderMerge.Add(p);
+                    else if (source.ExistsAsDirectory && !dest.IsValidAsDirectory(source.CompletePath)) invalidPath.Add(p);
+                    else if (source.ExistsAsFile && source.IsFileLocked) inUse.Add(p);
+                    else if (source.ExistsAsFile || source.ExistsAsDirectory) ok.Add(p);
+                    else throw new Exception("The path is nor a file or a folder.");
                 }
                 catch (Exception ex)
                 {
@@ -63,8 +55,32 @@ namespace BatchRenamerExtension
                 }
             }
 
-            
-            foreach (var p in renames)
+            if (invalidPath.Count > 0)
+            {
+                var desc = String.Join("\n", invalidPath.Select(x => x.Item1));
+                ShowError("Invalid path value", "This path(s) are invalid:\n\n" + desc);
+                return Tuple.Create(success, false);
+            }
+            if (folderMerge.Count > 0)
+            {
+                var desc = String.Join("\n", folderMerge.Select(x => x.Item2 + " -> " + x.Item1));
+                ShowError("Directory merge", "Directory merge is unsupported\n\n" + desc);
+                return Tuple.Create(success, false);
+            }
+            if (inUse.Count > 0)
+            {
+                var desc = String.Join("\n", inUse.Select(x => x.Item2));
+                ShowError("Files in use", "This file(s) are in use:\n\n" + desc);
+                return Tuple.Create(success, false);
+            }
+            if (overwriting.Count > 0)
+            {
+                var desc = String.Join("\n", overwriting.Select(x => x.Item1));
+                if (!Ask("Overwrite files?", "Do you want to overwrite these files:\n\n" + desc)) return Tuple.Create(success, false);
+                ok.AddRange(overwriting);
+            }
+
+            foreach (var p in ok)
             {
                 var source = p.Item2;
                 var dest = p.Item1;
@@ -72,19 +88,12 @@ namespace BatchRenamerExtension
                 {
                     if (source.ExistsAsFile)
                     {
-                        if(dest.ExistsAsFile)
-                        {
-                            File.Delete(dest.CompletePath);
-                        }
+                        if (dest.ExistsAsFile) File.Delete(dest.CompletePath);
                         File.Move(source.CompletePath, dest.CompletePath);
                     }
                     else if (source.ExistsAsDirectory)
                     {
                         Directory.Move(source.CompletePath, dest.CompletePath);
-                    }
-                    else
-                    {
-                        throw new Exception("The path is nor a file or a folder.");
                     }
                     success.Add(source);
                 }
